4.58 Implemented numberInStock()
Used a placeholder variable that calls findProduct() to locate a product with the given ID number. If a product is founds, the method then calls the getQuantity method and returns that int.
If no product is found and findProduct returns null, numberInStock returns 0.
Completed 4.58
(4.57 was completed in the last commit. I forgot to include that comment)